+++ packages/map-template/src/atoms/isLocationClickedState.js
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+const isLocationClickedState = atom({
+    key: 'isLocationClicked',
+    default: false
+});
+
+export default isLocationClickedState;