+++ packages/map-template/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.17.0] - 2023-11-06
+
+### Added
+
+- Added support for filter, search, highlight features from the SDK.
+- Implemented functionality for selecting a location and centering it on the map.
+
+### Fixed
+
+- Fixed floor selector not updating when exiting directions.
+- Fixed floor selector not updating when selecting a location on a different floor. 
+
 ## [1.16.0] - 2023-10-09
 
 ### Added